Fix scoring, which was ballsed up
My previous performance fix, which attempted to accumulate faced
statistics slowly over time, turned out to be rubish. This is because
you need the past data to recognize that pairings have happened before.
So the result of applying the calculations to each match was "Everyone
faces each other once", and no further data was available. Which was
silly.

So instead, build match data into the existing collection of
faced-statistics, calc scores, then remove the data about the matches we
just analyzed. This avoids deep copying the previous schedule score
dict, and is still vaguely fast.