Strongly encourage balancing min-gap counts
Previously the search was allowed to bunch up gaps for one team
if this reduced gaps for more than one other team. Adjusting the
cost of the Nth gap of a given size discourages this. It's still
not perfect as this may create imbalances between e.g: the third
gap of size three versus the second of size two, but it's definitely
better than before.